+++ src/shared/iengine.h
@@ -1,11 +1,11 @@
 // the interface the game uses to access the engine
 
-extern int curtime;          // scaled duration since last step
-extern int lastmillis;       // totalmillis value at last step
-extern int elapsedtime;      // wall clock duration since last step
-extern int totalmillis;      // total elapsed wall clock duration since engine start
-extern int curframetime;     // scaled duration since last frame
-extern int starttime;        // time from start
+extern Uint64 curtime;          // scaled duration since last step
+extern Uint64 lastmillis;       // totalmillis value at last step
+extern Uint64 elapsedtime;      // wall clock duration since last step
+extern Uint64 totalmillis;      // total elapsed wall clock duration since engine start
+extern Uint64 curframetime;     // scaled duration since last frame
+extern Uint64 starttime;        // time from start
 extern uint totalsecs;
 extern int gamespeed, paused;
 extern time_t walltime;      // wall time at engine startup (add totalmillis/1000 for current wall time)
@@ -422,7 +422,8 @@ enum
 	SND_NO_ALT  = 1<<1,
 	SND_USE_ALT = 1<<2
 };
-
+// TODO: SDL3_mixer
+/*
 extern int playsound(int n, const vec *loc = NULL, extentity *ent = NULL, int flags = 0, int loops = 0, int fade = 0, int chanid = -1, int radius = 0, int expire = -1);
 extern int playsoundname(const char *s, const vec *loc = NULL, int vol = 0, int flags = 0, int loops = 0, int fade = 0, int chanid = -1, int radius = 0, int expire = -1);
 extern void preloadsound(int n);
@@ -431,6 +432,7 @@ extern bool stopsound(int n, int chanid, int fade = 0);
 extern void stopsounds();
 extern void initsound();
 extern void resetsound();
+*/
 
 // rendermodel
 enum { MDL_CULL_VFC = 1<<0, MDL_CULL_DIST = 1<<1, MDL_CULL_OCCLUDED = 1<<2, MDL_CULL_QUERY = 1<<3, MDL_SHADOW = 1<<4, MDL_DYNSHADOW = 1<<5, MDL_LIGHT = 1<<6, MDL_DYNLIGHT = 1<<7, MDL_FULLBRIGHT = 1<<8, MDL_NORENDER = 1<<9, MDL_LIGHT_FAST = 1<<10, MDL_HUD = 1<<11, MDL_GHOST = 1<<12 };
